+++ modules/tile-converter/src/i3s-converter/helpers/gltf-attributes.ts
@@ -1,33 +1,27 @@
-import type {B3DMContent} from '@loaders.gl/3d-tiles';
+import type {Tiles3DTileContent} from '@loaders.gl/3d-tiles';
 import type {GLTFAccessorPostprocessed, GLTFNodePostprocessed} from '@loaders.gl/gltf';
 import type {B3DMAttributesData} from '../../i3s-attributes-worker';
+import {Matrix4, Vector3} from '@math.gl/core';
+import {BoundingSphere, OrientedBoundingBox} from '@math.gl/culling';
+import {Ellipsoid} from '@math.gl/geospatial';
 
 type AttributesObject = {
   [k: string]: GLTFAccessorPostprocessed;
 };
 
 /**
- * Keep only values for B3DM attributes to pass data to worker thread.
- * @param attributes
- */
-function getB3DMAttributesWithoutBufferView(attributes: AttributesObject): AttributesObject {
-  const attributesWithoutBufferView = {};
-
-  for (const attributeName in attributes) {
-    attributesWithoutBufferView[attributeName] = {
-      value: attributes[attributeName].value
-    };
-  }
-
-  return attributesWithoutBufferView;
-}
-
-/**
  * Prepare attributes for conversion to avoid binary data breaking in worker thread.
- * @param tileContent
+ * @param tileContent - 3DTiles tile content
+ * @param tileTransform - transformation matrix of the tile, calculated recursively multiplying
+ *                        transform of all parent tiles and transform of the current tile
+ * @param boundingVolume - initialized bounding volume of the source tile
  * @returns
  */
-export function prepareDataForAttributesConversion(tileContent: B3DMContent): B3DMAttributesData {
+export function prepareDataForAttributesConversion(
+  tileContent: Tiles3DTileContent,
+  tileTransform: Matrix4,
+  boundingVolume: OrientedBoundingBox | BoundingSphere
+): B3DMAttributesData {
   let nodes =
     tileContent.gltf?.scene?.nodes ||
     tileContent.gltf?.scenes?.[0]?.nodes ||
@@ -56,8 +50,11 @@ export function prepareDataForAttributesConversion(tileContent: B3DMContent): B3
 
   prepareNodes(nodes);
 
-  const cartographicOrigin = tileContent.cartographicOrigin;
-  const cartesianModelMatrix = tileContent.cartesianModelMatrix;
+  const {cartographicOrigin, modelMatrix: cartesianModelMatrix} = calculateTransformProps(
+    tileContent,
+    tileTransform,
+    boundingVolume
+  );
 
   return {
     nodes,
@@ -65,6 +62,72 @@ export function prepareDataForAttributesConversion(tileContent: B3DMContent): B3
     cartographicOrigin,
     cartesianModelMatrix
   };
+}
+
+/**
+ * Keep only values for glTF attributes to pass data to worker thread.
+ * @param attributes - geometry attributes
+ */
+function getB3DMAttributesWithoutBufferView(attributes: AttributesObject): AttributesObject {
+  const attributesWithoutBufferView = {};
+
+  for (const attributeName in attributes) {
+    attributesWithoutBufferView[attributeName] = {
+      value: attributes[attributeName].value
+    };
+  }
+
+  return attributesWithoutBufferView;
+}
+
+/**
+ * Calculate transformation properties to transform vertex attributes (POSITION, NORMAL, etc.)
+ * from METER_OFFSET coorditantes to LNGLAT_OFFSET coordinates
+ * @param tileContent - 3DTiles tile content
+ * @param tileTransform - transformation matrix of the tile, calculated recursively multiplying
+ *                        transform of all parent tiles and transform of the current tile
+ * @param boundingVolume - initialized bounding volume of the source tile
+ * @returns modelMatrix - transformation matrix to transform coordinates to cartographic coordinates
+ *          cartographicOrigin - tile origin coordinates to calculate offsets
+ */
+export function calculateTransformProps(
+  tileContent: Tiles3DTileContent,
+  tileTransform: Matrix4,
+  boundingVolume: OrientedBoundingBox | BoundingSphere
+): {modelMatrix: Matrix4; cartographicOrigin: Vector3} {
+  const {rtcCenter, gltfUpAxis} = tileContent;
+  const {center} = boundingVolume;
+
+  let modelMatrix = new Matrix4(tileTransform);
+
+  // Translate if appropriate
+  if (rtcCenter) {
+    modelMatrix.translate(rtcCenter);
+  }
+
+  // glTF models need to be rotated from Y to Z up
+  // https://github.com/AnalyticalGraphicsInc/3d-tiles/tree/master/specification#y-up-to-z-up
+  switch (gltfUpAxis) {
+    case 'Z':
+      break;
+    case 'Y':
+      const rotationY = new Matrix4().rotateX(Math.PI / 2);
+      modelMatrix = modelMatrix.multiplyRight(rotationY);
+      break;
+    case 'X':
+      const rotationX = new Matrix4().rotateY(-Math.PI / 2);
+      modelMatrix = modelMatrix.multiplyRight(rotationX);
+      break;
+    default:
+      break;
+  }
+
+  const cartesianOrigin = new Vector3(center);
+  const cartographicOrigin = Ellipsoid.WGS84.cartesianToCartographic(
+    cartesianOrigin,
+    new Vector3()
+  );
+  return {modelMatrix, cartographicOrigin};
 }
 
 /**